rename exline folder to distil
post-rebase clean up
