+++ distil/primitives/missing_indicator.py
@@ -10,7 +10,7 @@ import numpy as np
 
 from sklearn.impute import MissingIndicator
 
-from exline.primitives.utils import MISSING_VALUE_INDICATOR
+from distil.primitives.utils import MISSING_VALUE_INDICATOR
 
 logger = logging.getLogger(__name__)
 
@@ -37,9 +37,9 @@ class MissingIndicatorPrimitive(unsupervised_learning.UnsupervisedLearnerPrimiti
             'id': '15587104-0e81-4970-add3-668da63be95b',
             'version': '0.1.0',
             'name': "Missing indicator",
-            'python_path': 'd3m.primitives.data_transformation.missing_indicator.ExlineMissingIndicator',
+            'python_path': 'd3m.primitives.data_transformation.missing_indicator.DistilMissingIndicator',
             'source': {
-                'name': 'exline',
+                'name': 'Distil',
                 'contact': 'mailto:cbethune@uncharted.software',
                 'uris': [
                     'https://github.com/uncharted-distil/distil-primitives/primitives/missing_indicator.py',
@@ -48,7 +48,7 @@ class MissingIndicatorPrimitive(unsupervised_learning.UnsupervisedLearnerPrimiti
             },
             'installation': [{
                 'type': metadata_base.PrimitiveInstallationType.PIP,
-                'package_uri': 'git+https://github.com/uncharted-distil/distil-primitives.git@{git_commit}#egg=d3m-exline'.format(
+                'package_uri': 'git+https://github.com/uncharted-distil/distil-primitives.git@{git_commit}#egg=distil-primitives'.format(
                     git_commit=d3m_utils.current_git_commit(os.path.dirname(__file__)),
                 ),
             }],
